add workflow that checks for resource updates and remove auto updating resources
