+++ .github/workflows/check-for-resources-update.yml
@@ -1,0 +1,23 @@
+name: Check for utils resources update
+on:
+  schedule:
+    - cron: 0 0 * * 0
+
+jobs:
+  check-for-resources-update:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Install Node.js
+        uses: actions/setup-node@v3
+        with:
+          node-version: 18
+      - name: Install Packages
+        run: npm install
+      - name: Update
+        run: npm run update-resources
+      - name: Check changes
+        run: |
+            git add --all && \
+            git diff-index --cached HEAD --stat --exit-code
