+++ tools/update.js
@@ -11,8 +11,3 @@ require('./update-lib-flat-configs')
 require('./update-lib-index')
 require('./update-docs')
 require('./update-docs-rules-index')
-
-if (process.env.IN_VERSION_SCRIPT) {
-  require('./update-vue3-export-names')
-  require('./update-resources')
-}
